fix: set annotations for broadcast info
We are currently JSON-encoding the broadcast details as a proposed
header annotation, so that upon receiving a proposed header, the
recipient has all the information they need to accept the broadcast.

In the future we will use something more efficient than JSON to encode
this, but at this point in testing it suffices.

The network compliance tests involving only proposed headers are passing
at this point, but the others are still failing.