feat: add unaggregated vote state
The statefulness of dragon wingspan connections doesn't line up great
with the prior stateless expectations of tmp2p, but this should cover
the state implementation for unaggregated (i.e. ed25519) signatures. We
will get this wired up with the tmp2p.Connection implementation before
working on the aggregated signature state.