feat: begin integrating wingspan unaggregated votes
The code compiles but this does not yet include tests.

The interface on the state implementations effectively fills in the
blanks for the wingspan protocol; but, it is currently difficult to set
up "nodes" for the protocol in a test. So next, I am going to evaluate
whether we can use existing utilities from dragon for testing this, or
if we need to expose some types that are currently marked internal.

I think the ListenerSet is the type we want from dragon, but that is
internal and it uses a few other internal types, if I recall correctly.