feat: add initial sketch of NetworkAdapter
This type begins integrating the breathcast adapter and a wingspan
protocol instance, for a unified collection of all the necessary message
types for the core Gordian p2p layer.

This does not yet answer where QUIC streams are accepted. That could
happen inside or outside the NetworkAdapter. I'm leaning towards inside,
so that wingspan and breathcast streams can be handled directly, and
application-layer streams can be exposed directly to the application
without having to check if they belong to one of the other driver-level
protocols.

This type is still incomplete and missing some tests, but this seems
like a good checkpoint before getting into the stream handling code.