feat: partially implement streamAccepter
As part of the network adapter, we need dedicated goroutines for each
connection, attempting to accept both bidirectional and unidirectional
streams. Upon accepting a stream, we check the protocol ID to see if it
matches the configured protocols. If it does match, we consider whether
it is in the active set of sessions. If it doesn't match, expose the
stream and the protocol ID to the application layer, and they can decide
whether to match the stream to an existing application level protocol or
to close the connection.