feat: broadcast to directly connected peer
This is missing datagram handling, which probably needs to happen in the
StreamAccepter; at which point that type may be misnamed. But it still
works because of the synchronous fallback.

And it appears that the broadcast isn't making the hop from node 1 to
node 2 in test. Without debugging, I'm not yet sure if that is due to
the stub implementation of handleBreathcastCheck or something else.