feat: handle application-layer datagrams
This was originally going to continue to include managing breathcast
packet datagrams, but I ran into a mismatch in how breathcast's
broadcast IDs are used and how the network adapter stores its sessions.
The session store maps by block hash, but the broadcast ID uses the
proposer index.

So, we will commit here and then sort out the broadcast ID keying for
session management, and then return to breathcast datagrams.