chore: add loggers to gdwsu state types
In attempting to get the integration tests passing, it is fairly helpful
to be able to trace how messages are flowing between nodes. We may or
may not keep the debug logs around, but having the loggers available
seems worthwhile.

A missing piece here is that one central state instance is expected to
have many inbound and outbound derived states, and at the moment we are
unable to distinguish the derived instances due to having no connection
context available at instantiation. Maybe that's a future
dragon/wingspan API change.