test: add skipped gordian integration test
The tests compile but all three integration apps currently fail.

The simple two without block data are failing due to what I assume is a
trivial in dragon. The full block data in those apps is able to fit in a
single packet, and when we attempt to reconstruct at that point, the
code panics because we are not supposed to attempt reconstruction when
we have the full data. There are at least a couple ways to handle that
case.

But more interesting is the random data case where we transfer 32kb
blocks of randomly generated bytes. We seem to regularly get up to
height 2 or 3 when one or two validators get out of sync, and then they
don't output a precommit. I haven't yet determined the root cause on
this. The daisy chain p2p layer in the gordian repo reliably passes this
test, so it's likely something in dragon. Still investigating.