fix: handle "unexpected" vote for next height
Prior to this change, we dropped votes for the next height before we
received an "activate" signal from the engine. Therefore, integration
tests failed quickly as there are necessarily going to be timing
differences between each node finishing their commit delay; the chain
would inevitably halt as validators got "stuck" due to inadvertently
rejecting vote streams.

Catching up perhaps would have partially addressed this, but it really
would have just masked the problem.

Now, if we see a vote for the next height, and if we have a proposed
header for the current height, we can use the NextValidators field on
the header to confirm that the votes have the right set of public keys,
and therefore the votes can be verified and accepted.

This may have introduced some flakiness in the gdna tests; I will look
at that next.