fix: handle context cancellation gracefully in more cases
The test flakiness was due to one network adapter not being informed of
the first network view update, and then fixing that highlighted a couple
remaining TODOs on error handling, which were being triggered by context
cancellation. The non-context errors will still need to be handled more
gracefully in the future.