chore: require one network view update before main loop
This is setting up structure for using the network view updates' round
view's validator set as the source of truth, instead of the current
approach of inspecting proposed headers. The proposed headers could be
missing at the time we receive a vote, for instance.

We will still need to fall back to proposed headers in certain missing
height situations, but that should be the exception.